requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
dj-stripe>=2.8.0
channels>=4.0.0
celery>=5.3.0
django-allauth>=0.57.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
google-api-python-client>=2.108.0
gunicorn>=21.2.0
whitenoise>=6.5.0
daphne>=4.0.0 